Tutorial fix, maybe?
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3957 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/format.cc
@@ -346,7 +346,7 @@ std::string formatted_string::to_colour_string() const
         if (ops[i] == FSOP_TEXT)
         {
             // gotta double up those '<' chars ...
-            unsigned int start = st.size();
+            size_t start = st.size();
             st += ops[i].text;
             
             while (true)
